sign up pages and password relateds

+++ src/pages/sign-up/sign-up-user-email/sign-up-user-email.shared.tsx
@@ -1,25 +1,31 @@
+import { useNavigate } from 'react-router-dom';
+import { register, preRegister, RegisterReq, handleError } from 'src/core/api';
 import { useForm } from 'src/core/form';
 import { getFormValues } from 'src/core/form/customValidators/formValues';
-import { handleError } from 'src/core/http';
 
 import { formModel } from './sign-up-user-email.form';
-import { preRegister, register } from './sign-up-user-email.services';
 
 export const useSignUpUserEmailShared = () => {
   const form = useForm(formModel);
-  const navigate = {};
+  const navigate = useNavigate();
 
   async function onSubmit() {
-    const formValues = getFormValues(form);
-    await preRegister({ email: form.controls.email.value });
+    const formValues = getFormValues(form) as RegisterReq;
+
+    const response = await preRegister({ email: formValues.email });
+    if (response.email === 'EXIST') {
+      handleError(Error('email already exists'));
+      return;
+    }
+
     register(formValues)
       .then(() => localStorage.setItem('email', formValues.email))
-      .then(() => navigate({ to: '../verification' }))
+      .then(() => navigate('../verification'))
       .catch(handleError());
   }
 
   function navigateToSignIn() {
-    navigate({ to: '/sign-in' });
+    navigate('/sign-in');
   }
 
   return { form, navigateToSignIn, onSubmit };
